*** empty log message ***
git-svn-id: http://qboxcode.org/svn/qb/trunk@536 cba15fb0-1239-40c8-b417-11db7ca47a34

+++ src/testFourierTransform.C
@@ -17,7 +17,7 @@ using namespace std;
 
 int fft_flops(int n)
 {
-  return 5.0 * n * log((double) n) / log(2.0);
+  return (int) (5.0 * n * log((double) n) / log(2.0));
 }
 
 int main(int argc, char **argv)
